requirements.txt
Django==2.2
django-cors-headers==3.0.1
djangorestframework==3.9.2
djangorestframework-simplejwt==4.3.0
mysqlclient==1.4.2.post1
PyJWT==1.7.1
pytz==2018.9
sqlparse==0.3.0
